hospitalv3/backend/requirements.txt
django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
drf-spectacular==0.29.0
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
gunicorn==23.0.0
whitenoise==6.8.2
